+++ mirtop/mirna/mapper.py
@@ -56,33 +56,16 @@ def read_gtf_to_mirna(gtf):  # deprecate in the future
     """
     if not gtf:
         return gtf
-    db = defaultdict(list)
-    db_mir = defaultdict(dict)
-    id_dict = dict()
-    with open(gtf) as in_handle:
-        for line in in_handle:
-            if line.startswith("#"):
-                continue
-            cols = line.strip().split("\t")
-            logger.debug("MAP:: line:%s" % cols)
-            name = [n.split("=")[1] for n in cols[-1].split(";")
-                    if n.startswith("Name")]
-            idname = [n.split("=")[1] for n in cols[-1].split(";")
-                      if n.startswith("ID")]
-            chrom, start, end, strand = cols[0], cols[3], cols[4], cols[6]
-            logger.debug("MAP:: idname:%s" % idname)
-            logger.debug("MAP:: name:%s" % name)
-            id_dict[idname[0]] = name[0] # MIMA to sps-Y-X-5/3p
-            if cols[2] == "miRNA_primary_transcript":
-                db[idname[0]] = [chrom, int(start), int(end), strand]
-            if cols[2] == "miRNA":
-                parent = [n.split("=")[1] for n in cols[-1].split(";")
-                          if n.startswith("Derives_from")]
-                db_mir[name[0]].update({id_dict[parent[0]]: db[parent[0]]})
-                logger.debug("MAP:: mirna:%s" % name[0])
-                logger.debug("MAP:: precursor:%s" % id_dict[parent[0]])
-                logger.debug("MAP:: precursor pos %s" % db[parent[0]])
-    return db_mir
+    if _guess_database_file(gtf).find("miRBase") > -1:
+        mapped = read_gtf_to_precursor_mirbase(gtf, primary=True)
+    elif _guess_database_file(gtf).find("MirGeneDB") > -1:
+        mapped = read_gtf_to_precursor_mirgenedb(gtf, primary=True)
+    else:
+        logger.info("Database different than miRBase or MirGeneDB")
+        logger.info("If you get an error when loading,")
+        logger.info("report it to https://github.com/miRTop/mirtop/issues")
+        mapped = read_gtf_to_precursor_mirbase(gtf)
+    return mapped
 
 
 def read_gtf_chr2mirna(gtf):  # read from read_gtf_to_precursor
@@ -163,7 +146,8 @@ def read_gtf_to_precursor(gtf):  # need to adapt to mirgenedb
     return mapped
 
 
-def _parse_db_mir(db_mir, db, id_dict, map_dict):
+def _parse_db_mir(db_mir, db, id_dict):
+    map_dict = defaultdict(dict)
     for mir in db_mir:
         parent = db_mir[mir][4]
         precursor = db[id_dict[parent]]
@@ -193,7 +177,29 @@ def _parse_db_mir(db_mir, db, id_dict, map_dict):
     return map_dict
 
 
-def read_gtf_to_precursor_mirgenedb(gtf):
+def _parse_db_mir_genomic(db_mir, db, id_dict):
+    map_dict = defaultdict(dict)
+    for mir in db_mir:
+        parent = db_mir[mir][4]
+        precursor = db[id_dict[parent]]
+        logger.debug("MAP::%s %s %s" % (
+            id_dict[parent], precursor[1], precursor[2]))
+        logger.debug("MAP::%s %s %s" % (mir, db_mir[mir][1], db_mir[mir][2]))
+        if precursor[3] != db_mir[mir][3]:
+            logger.warning("%s -> %s" % (id_dict[parent], mir))
+            logger.warning("miRNA strand doesn't match with precursor strand:"
+                           " %s - %s" % (db_mir[mir][3], precursor[3]))
+            next
+        if precursor[0] != db_mir[mir][0]:
+            logger.warning("%s -> %s" % (id_dict[parent], mir))
+            logger.warning("miRNA chr doesn't match with precursor chr:"
+                           " %s - %s" % (db_mir[mir][0], precursor[0]))
+            next
+        map_dict[mir[1]].update({id_dict[parent]: db_mir[mir]})
+    return map_dict
+
+
+def read_gtf_to_precursor_mirgenedb(gtf, primary=False):
     """
     Load GTF file with precursor positions on genome
     Return dict with key being precursor name and
@@ -214,7 +220,6 @@ def read_gtf_to_precursor_mirgenedb(gtf):
     db = defaultdict(list)
     db_mir = defaultdict(list)
     id_dict = dict()
-    map_dict = defaultdict(dict)
     with open(gtf) as in_handle:
         for line in in_handle:
             if line.startswith("#"):
@@ -236,10 +241,12 @@ def read_gtf_to_precursor_mirgenedb(gtf):
                                           strand, parent]
                 logger.debug("MAP:: mirna:%s" % name)
                 logger.debug("MAP:: pos %s" % db_mir[(parent, name)])
-    return _parse_db_mir(db_mir, db, id_dict, map_dict)
+    if primary:
+        return _parse_db_mir_genomic(db_mir, db, id_dict)
+    return _parse_db_mir(db_mir, db, id_dict)
 
 
-def read_gtf_to_precursor_mirbase(gtf):
+def read_gtf_to_precursor_mirbase(gtf, primary=False):
     """
     Load GTF file with precursor positions on genome
     Return dict with key being precursor name and
@@ -260,7 +267,6 @@ def read_gtf_to_precursor_mirbase(gtf):
     db = defaultdict(list)
     db_mir = defaultdict(list)
     id_dict = dict()
-    map_dict = defaultdict(dict)
     with open(gtf) as in_handle:
         for line in in_handle:
             if line.startswith("#"):
@@ -282,7 +288,9 @@ def read_gtf_to_precursor_mirbase(gtf):
                                                 strand, parent[0]]
                 logger.debug("MAP:: mirna:%s" % name[0])
                 logger.debug("MAP:: pos %s" % db_mir[(parent[0], name[0])])
-    return _parse_db_mir(db_mir, db, id_dict, map_dict)
+    if primary:
+        return _parse_db_mir_genomic(db_mir, db, id_dict)
+    return _parse_db_mir(db_mir, db, id_dict)
 
 
 def liftover_genomic_precursor(read, genome, hairpin, expected=None):